Move the info cache creation to the saga feature where it belongs
